[REV] purchase, account, sale: add account.order(.line).mixin
Reverted because:

- Mixin added more complexity than simplification (i.e. more lines added
  than removed)
- Downpayment functionality for POs as is in this commit is not desired
- POs and SOs should be allowed to evolve independently of each other
  and this mixin restricts that

Because subsequent renaming was done on top of this mixin, we revert
those renamings as well to avoid painful conflicts

This reverts commit 0257fcb9cfef5bac3c267f2e85cdc04d3701c4f1.

Original task: 3600910

X-original-commit: 465da9d7e63157a312ea83b12b760520dfaf59e6
