+++ addons/account/models/__init__.py
@@ -6,8 +6,6 @@ from . import ir_http
 from . import res_partner_bank
 from . import account_account_tag
 from . import account_account
-from . import account_order_line_mixin
-from . import account_order_mixin
 from . import account_journal
 from . import account_tax
 from . import account_reconcile_model
