+++ addons/account/models/account_move_line.py
@@ -418,7 +418,6 @@ class AccountMoveLine(models.Model):
              "associated partner",
     )
     is_refund = fields.Boolean(compute='_compute_is_refund')
-    is_downpayment = fields.Boolean()
 
     _sql_constraints = [
         (
@@ -1186,9 +1185,6 @@ class AccountMoveLine(models.Model):
             'target': 'new',
             'type': 'ir.actions.act_window',
         }
-
-    def _get_order_lines(self):
-        return self.env['account.order.line.mixin']
 
     # -------------------------------------------------------------------------
     # INVERSE METHODS
