+++ addons/account/views/account_move_views.xml
@@ -1052,7 +1052,6 @@
                                         <field name="product_id"
                                                optional="show"
                                                widget="product_label_section_and_note_field"
-                                               readonly="is_downpayment"
                                                domain="
                                                     context.get('default_move_type') in ('out_invoice', 'out_refund', 'out_receipt')
                                                     and [('sale_ok', '=', True)]
