+++ addons/purchase/__init__.py
@@ -5,4 +5,3 @@ from . import controllers
 from . import models
 from . import report
 from . import populate
-from . import wizard
