+++ addons/purchase/report/purchase_quotation_templates.xml
@@ -30,8 +30,7 @@
                     </tr>
                 </thead>
                 <tbody>
-                    <t t-set="lines_to_report" t-value="o._get_order_lines_to_report()"/>
-                    <t t-foreach="lines_to_report" t-as="order_line">
+                    <t t-foreach="o.order_line" t-as="order_line">
                         <tr t-att-class="'fw-bold o_line_section' if order_line.display_type == 'line_section' else 'fst-italic o_line_note' if order_line.display_type == 'line_note' else ''">
                             <t t-if="not order_line.display_type">
                                 <td id="product">
