+++ addons/purchase/security/ir.model.access.csv
@@ -37,4 +37,3 @@ access_account_account_purchase_manager,account.account purchase manager,account
 access_purchase_bill_union,access_purchase_bill_union,model_purchase_bill_union,purchase.group_purchase_user,1,0,0,0
 access_report_purchase_order,purchase.stock.report,model_purchase_report,purchase.group_purchase_manager,1,0,0,0
 access_report_purchase_order_user,purchase.stock.report user,model_purchase_report,purchase.group_purchase_user,1,0,0,0
-access_purchase_advance_payment_wizard,access.purchase.advance.payment.wizard,model_purchase_advance_payment_wizard,purchase.group_purchase_user,1,1,1,0
