+++ addons/purchase/security/purchase_security.xml
@@ -30,11 +30,6 @@
         <field name="category_id" ref="base.module_category_hidden"/>
     </record>
 
-    <record id="group_down_payment" model="res.groups">
-        <field name="name">Allow creating vendor bills for down payments on purchase orders</field>
-        <field name="category_id" ref="base.module_category_hidden"/>
-    </record>
-
 </data>
 <data noupdate="1">
     <record model="res.groups" id="base.group_user">
