+++ addons/purchase/tests/__init__.py
@@ -8,4 +8,3 @@ from . import test_access_rights
 from . import test_accrued_purchase_orders
 from . import test_purchase_tax_totals
 from . import test_purchase_dashboard
-from . import test_purchase_downpayment
