+++ addons/purchase_requisition/wizard/purchase_requisition_create_alternative.py
@@ -80,10 +80,7 @@ class PurchaseRequisitionCreateAlternative(models.TransientModel):
             'origin': self.origin_po_id.origin,
         }
         if self.copy_products and self.origin_po_id:
-            vals['order_line'] = [
-                Command.create(self._get_alternative_line_value(line))
-                for line in self.origin_po_id.order_line if not line.is_downpayment
-            ]
+            vals['order_line'] = [Command.create(self._get_alternative_line_value(line)) for line in self.origin_po_id.order_line]
         return vals
 
     @api.model
