+++ addons/sale/models/sale_order_line.py
@@ -7,16 +7,16 @@ from datetime import timedelta
 
 from markupsafe import Markup
 
-from odoo import _, api, fields, models
+from odoo import _, _lt, api, fields, models
 from odoo.exceptions import UserError
 from odoo.fields import Command
 from odoo.osv import expression
-from odoo.tools import float_compare, float_is_zero, groupby
+from odoo.tools import float_compare, float_is_zero, format_date, groupby
 
 
 class SaleOrderLine(models.Model):
     _name = 'sale.order.line'
-    _inherit = ['analytic.mixin', 'account.order.line.mixin']
+    _inherit = 'analytic.mixin'
     _description = "Sales Order Line"
     _rec_names_search = ['name', 'order_id.name']
     _order = 'order_id, sequence, id'
@@ -46,6 +46,10 @@ class SaleOrderLine(models.Model):
     company_id = fields.Many2one(
         related='order_id.company_id',
         store=True, index=True, precompute=True)
+    currency_id = fields.Many2one(
+        related='order_id.currency_id',
+        depends=['order_id.currency_id'],
+        store=True, precompute=True)
     order_partner_id = fields.Many2one(
         related='order_id.partner_id',
         string="Customer",
@@ -61,17 +65,29 @@ class SaleOrderLine(models.Model):
     tax_country_id = fields.Many2one(related='order_id.tax_country_id')
 
     # Fields specifying custom line logic
+    display_type = fields.Selection(
+        selection=[
+            ('line_section', "Section"),
+            ('line_note', "Note"),
+        ],
+        default=False)
     is_configurable_product = fields.Boolean(
         string="Is the product configurable?",
         related='product_template_id.has_configurable_attributes',
         depends=['product_id'])
+    is_downpayment = fields.Boolean(
+        string="Is a down payment",
+        help="Down payments are made when creating invoices from a sales order."
+            " They are not copied when duplicating a sales order.")
     is_expense = fields.Boolean(
         string="Is expense",
         help="Is true if the sales order line comes from an expense or a vendor bills")
 
     # Generic configuration fields
     product_id = fields.Many2one(
-        ondelete='restrict',
+        comodel_name='product.product',
+        string="Product",
+        change_default=True, ondelete='restrict', index='btree_not_null',
         domain="[('sale_ok', '=', True)]")
     product_template_id = fields.Many2one(
         string="Product Template",
@@ -102,14 +118,22 @@ class SaleOrderLine(models.Model):
         store=True, readonly=False, precompute=True, ondelete='restrict')
     is_product_archived = fields.Boolean(compute="_compute_is_product_archived")
 
+    name = fields.Text(
+        string="Description",
+        compute='_compute_name',
+        store=True, readonly=False, required=True, precompute=True)
+
     product_uom_qty = fields.Float(
         string="Quantity",
         compute='_compute_product_uom_qty',
         digits='Product Unit of Measure', default=1.0,
         store=True, readonly=False, required=True, precompute=True)
     product_uom = fields.Many2one(
+        comodel_name='uom.uom',
+        string="Unit of Measure",
         compute='_compute_product_uom',
-        store=True, readonly=False, precompute=True, ondelete='restrict')
+        store=True, readonly=False, precompute=True, ondelete='restrict',
+        domain="[('category_id', '=', product_uom_category_id)]")
     linked_line_id = fields.Many2one(
         string="Linked Order Line",
         comodel_name='sale.order.line',
@@ -124,23 +148,40 @@ class SaleOrderLine(models.Model):
 
     # Pricing fields
     tax_id = fields.Many2many(
+        comodel_name='account.tax',
+        string="Taxes",
         compute='_compute_tax_id',
-        store=True, readonly=False, precompute=True)
+        store=True, readonly=False, precompute=True,
+        context={'active_test': False},
+        check_company=True)
 
     # Tech field caching pricelist rule used for price & discount computation
     pricelist_item_id = fields.Many2one(
         comodel_name='product.pricelist.item',
         compute='_compute_pricelist_item_id')
 
-    price_unit = fields.Float(compute='_compute_price_unit', store=True, precompute=True)
+    price_unit = fields.Float(
+        string="Unit Price",
+        compute='_compute_price_unit',
+        digits='Product Price',
+        store=True, readonly=False, required=True, precompute=True)
+
+    discount = fields.Float(
+        string="Discount (%)",
+        compute='_compute_discount',
+        digits='Discount',
+        store=True, readonly=False, precompute=True)
 
     price_subtotal = fields.Monetary(
+        string="Subtotal",
+        compute='_compute_amount',
+        store=True, precompute=True)
+    price_tax = fields.Float(
+        string="Total Tax",
         compute='_compute_amount',
         store=True, precompute=True)
     price_total = fields.Monetary(
-        compute='_compute_amount',
-        store=True, precompute=True)
-    price_tax = fields.Monetary(
+        string="Total",
         compute='_compute_amount',
         store=True, precompute=True)
     price_reduce_taxexcl = fields.Monetary(
@@ -197,6 +238,11 @@ class SaleOrderLine(models.Model):
         compute='_compute_qty_invoiced',
         digits='Product Unit of Measure',
         store=True)
+    qty_to_invoice = fields.Float(
+        string="Quantity To Invoice",
+        compute='_compute_qty_to_invoice',
+        digits='Product Unit of Measure',
+        store=True)
 
     analytic_line_ids = fields.One2many(
         comodel_name='account.analytic.line', inverse_name='so_line',
@@ -251,17 +297,6 @@ class SaleOrderLine(models.Model):
                 name = f'{name} {additional_name}'
             so_line.display_name = name
 
-    def _additional_name_per_id(self):
-        return {
-            so_line.id: so_line._get_partner_display()
-            for so_line in self
-        }
-
-    def _get_partner_display(self):
-        self.ensure_one()
-        commercial_partner = self.order_partner_id.commercial_partner_id
-        return f'({commercial_partner.ref or commercial_partner.name})'
-
     @api.depends('product_id')
     def _compute_product_template_id(self):
         for line in self:
@@ -305,16 +340,20 @@ class SaleOrderLine(models.Model):
 
     @api.depends('product_id', 'linked_line_id', 'linked_line_ids')
     def _compute_name(self):
-        super()._compute_name()
         for line in self:
-            if not line.product_id:
+            if not line.product_id and not line.is_downpayment:
                 continue
 
             lang = line.order_id._get_lang()
             if lang != self.env.lang:
                 line = line.with_context(lang=lang)
 
-            line.name = line._get_sale_order_line_multiline_description_sale()
+            if line.product_id:
+                line.name = line._get_sale_order_line_multiline_description_sale()
+                continue
+
+            if line.is_downpayment:
+                line.name = line._get_downpayment_description()
 
     def _get_sale_order_line_multiline_description_sale(self):
         """ Compute a default multiline description for this sales order line.
@@ -378,6 +417,33 @@ class SaleOrderLine(models.Model):
         for patv in sorted_custom_ptav:
             pacv = self.product_custom_attribute_value_ids.filtered(lambda pcav: pcav.custom_product_template_attribute_value_id == patv)
             name += "\n" + pacv.display_name
+
+        return name
+
+    def _get_downpayment_description(self):
+        self.ensure_one()
+        if self.display_type:
+            return _("Down Payments")
+
+        dp_state = self._get_downpayment_state()
+        name = _lt("Down Payment")
+        if dp_state == 'draft':
+            name = _(
+                "Down Payment: %(date)s (Draft)",
+                date=format_date(self.env, self.create_date.date()),
+            )
+        elif dp_state == 'cancel':
+            name = _("Down Payment (Cancelled)")
+        else:
+            invoice = self._get_invoice_lines().filtered(
+                lambda aml: aml.quantity >= 0
+            ).move_id.filtered(lambda move: move.move_type == 'out_invoice')
+            if len(invoice) == 1 and invoice.payment_reference and invoice.invoice_date:
+                name = _(
+                    "Down Payment (ref: %(reference)s on %(date)s)",
+                    reference=invoice.payment_reference,
+                    date=format_date(self.env, invoice.invoice_date),
+                )
 
         return name
 
@@ -448,7 +514,6 @@ class SaleOrderLine(models.Model):
 
     @api.depends('product_id', 'product_uom', 'product_uom_qty')
     def _compute_price_unit(self):
-        super()._compute_price_unit()
         for line in self:
             # check if there is already invoiced amount. if so, the price shouldn't change as it might have been
             # manually edited
@@ -548,7 +613,6 @@ class SaleOrderLine(models.Model):
 
     @api.depends('product_id', 'product_uom', 'product_uom_qty')
     def _compute_discount(self):
-        super()._compute_discount()
         for line in self:
             if not line.product_id or line.display_type:
                 line.discount = 0.0
@@ -598,9 +662,25 @@ class SaleOrderLine(models.Model):
             **kwargs,
         )
 
-    @api.depends('product_uom_qty')
+    @api.depends('product_uom_qty', 'discount', 'price_unit', 'tax_id')
     def _compute_amount(self):
-        super()._compute_amount()
+        """
+        Compute the amounts of the SO line.
+        """
+        for line in self:
+            tax_results = self.env['account.tax']._compute_taxes(
+                [line._convert_to_tax_base_line_dict()],
+                line.company_id,
+            )
+            totals = list(tax_results['totals'].values())[0]
+            amount_untaxed = totals['amount_untaxed']
+            amount_tax = totals['amount_tax']
+
+            line.update({
+                'price_subtotal': amount_untaxed,
+                'price_tax': amount_tax,
+                'price_total': amount_untaxed + amount_tax,
+            })
 
     @api.depends('price_subtotal', 'product_uom_qty')
     def _compute_price_reduce_taxexcl(self):
@@ -673,6 +753,20 @@ class SaleOrderLine(models.Model):
         for so_line in lines_by_analytic:
             so_line.qty_delivered = mapping.get(so_line.id or so_line._origin.id, 0.0)
 
+    def _get_downpayment_state(self):
+        self.ensure_one()
+
+        if self.display_type:
+            return ''
+
+        invoice_lines = self._get_invoice_lines()
+        if all(line.parent_state == 'draft' for line in invoice_lines):
+            return 'draft'
+        if all(line.parent_state == 'cancel' for line in invoice_lines):
+            return 'cancel'
+
+        return ''
+
     def _get_delivered_quantity_by_analytic(self, additional_domain):
         """ Compute and write the delivered quantity of current SO lines, based on their related
             analytic lines.
@@ -725,6 +819,15 @@ class SaleOrderLine(models.Model):
                         qty_invoiced -= invoice_line.product_uom_id._compute_quantity(invoice_line.quantity, line.product_uom)
             line.qty_invoiced = qty_invoiced
 
+    def _get_invoice_lines(self):
+        self.ensure_one()
+        if self._context.get('accrual_entry_date'):
+            return self.invoice_lines.filtered(
+                lambda l: l.move_id.invoice_date and l.move_id.invoice_date <= self._context['accrual_entry_date']
+            )
+        else:
+            return self.invoice_lines
+
     # no trigger product_id.invoice_policy to avoid retroactively changing SO
     @api.depends('qty_invoiced', 'qty_delivered', 'product_uom_qty', 'state')
     def _compute_qty_to_invoice(self):
@@ -732,7 +835,6 @@ class SaleOrderLine(models.Model):
         Compute the quantity to invoice. If the invoice policy is order, the quantity to invoice is
         calculated from the ordered quantity. Otherwise, the quantity delivered is used.
         """
-        super()._compute_qty_to_invoice()
         for line in self:
             if line.state == 'sale' and not line.display_type:
                 if line.product_id.invoice_policy == 'order':
@@ -933,7 +1035,7 @@ class SaleOrderLine(models.Model):
                 vals['product_uom_qty'] = 0.0
 
         lines = super().create(vals_list)
-        if self.env.context.get('no_log_for_new_lines'):
+        if self.env.context.get('sale_no_log_for_new_lines'):
             return lines
 
         for line in lines:
@@ -1054,8 +1156,6 @@ class SaleOrderLine(models.Model):
     def compute_uom_qty(self, new_qty, stock_move, rounding=True):
         return self.product_uom._compute_quantity(new_qty, stock_move.product_uom, rounding)
 
-    # INVOICING #
-
     def _get_invoice_line_sequence(self, new=0, old=0):
         """
         Method intended to be overridden in third-party module if we want to prevent the resequencing
@@ -1074,7 +1174,7 @@ class SaleOrderLine(models.Model):
         :param optional_values: any parameter that should be added to the returned invoice line
         :rtype: dict
         """
-        res = super()._prepare_invoice_line(**optional_values)
+        self.ensure_one()
 
         # Compatibility fix for creating invoices from a SO since the computation of the line name has been changed in the account module.
         # Has to be removed as soon as the new behavior for the line name has been implemented in the sale module.
@@ -1083,14 +1183,19 @@ class SaleOrderLine(models.Model):
             line_name = re.sub(re.escape(self.product_id.display_name), '', line_name)
             line_name = re.sub(r'^\n', '', line_name)
             line_name = re.sub(r'(?<=\n) ', '', line_name)
-
-        res.update({
+        res = {
+            'display_type': self.display_type or 'product',
             'sequence': self.sequence,
             'name': line_name,
+            'product_id': self.product_id.id,
+            'product_uom_id': self.product_uom.id,
+            'quantity': self.qty_to_invoice,
+            'discount': self.discount,
             'price_unit': self.price_unit,
             'tax_ids': [Command.set(self.tax_id.ids)],
             'sale_line_ids': [Command.link(self.id)],
-        })
+            'is_downpayment': self.is_downpayment,
+        }
         self._set_analytic_distribution(res, **optional_values)
         if self.is_downpayment:
             res['account_id'] = self.invoice_lines.filtered('is_downpayment').account_id[:1].id
@@ -1111,21 +1216,33 @@ class SaleOrderLine(models.Model):
             else:
                 inv_line_vals['analytic_distribution'] = {analytic_account_id: 100}
 
+    def _prepare_procurement_values(self, group_id=False):
+        """ Prepare specific key for moves or other components that will be created from a stock rule
+        coming from a sale order line. This method could be override in order to add other custom key that could
+        be used in move/po creation.
+        """
+        return {}
+
     def _validate_analytic_distribution(self):
-        for line in self.filtered(
-            lambda sol: not sol.display_type and sol.state in ['draft', 'sent']
-        ):
-            line._validate_distribution(
-                product=line.product_id.id,
-                business_domain='sale_order',
-                company_id=line.company_id.id,
-            )
+        for line in self.filtered(lambda l: not l.display_type and l.state in ['draft', 'sent']):
+            line._validate_distribution(**{
+                'product': line.product_id.id,
+                'business_domain': 'sale_order',
+                'company_id': line.company_id.id,
+            })
 
-    def has_valued_move_ids(self):
-        return self.move_ids
+    #=== CORE METHODS OVERRIDES ===#
 
-    def _has_valid_qty_to_invoice(self, final=False):
-        return self.qty_to_invoice > 0 or (self.qty_to_invoice < 0 and final)
+    def _get_partner_display(self):
+        self.ensure_one()
+        commercial_partner = self.order_partner_id.commercial_partner_id
+        return f'({commercial_partner.ref or commercial_partner.name})'
+
+    def _additional_name_per_id(self):
+        return {
+            so_line.id: so_line._get_partner_display()
+            for so_line in self
+        }
 
     #=== HOOKS ===#
 
@@ -1202,13 +1319,6 @@ class SaleOrderLine(models.Model):
                 # price will be computed in batch with pricelist utils so not given here
             }
 
-    def _prepare_procurement_values(self, group_id=False):
-        """ Prepare specific key for moves or other components that will be created from a stock rule
-        coming from a sale order line. This method could be override in order to add other custom key that could
-        be used in move/po creation.
-        """
-        return {}
-
     #=== TOOLING ===#
 
     def _convert_to_sol_currency(self, amount, currency):
@@ -1233,3 +1343,6 @@ class SaleOrderLine(models.Model):
                 round=False,
             )
         return amount
+
+    def has_valued_move_ids(self):
+        return self.move_ids
