+++ addons/sale/tests/test_sale_order_down_payment.py
@@ -140,7 +140,7 @@ class TestSaleOrderDownPayment(TestSaleCommon):
 
         # Full Invoice
         invoicing_wizard = self.env['sale.advance.payment.inv'].create({
-            'order_ids': [Command.link(self.sale_order.id)],
+            'sale_order_ids': [Command.link(self.sale_order.id)],
             'advance_payment_method': 'delivered',
         })
         action = invoicing_wizard.create_invoices()
@@ -1033,7 +1033,7 @@ class TestSaleOrderDownPayment(TestSaleCommon):
         invoicing_wizard = self.env['sale.advance.payment.inv'].create({
             'advance_payment_method': 'fixed',
             'fixed_amount': sale_order.amount_total / 2.0,
-            'order_ids': [Command.link(sale_order.id)],
+            'sale_order_ids': [Command.link(sale_order.id)],
         })
 
         # Down payment invoice
@@ -1049,7 +1049,7 @@ class TestSaleOrderDownPayment(TestSaleCommon):
 
         # Full Invoice
         invoicing_wizard = self.env['sale.advance.payment.inv'].create({
-            'order_ids': [Command.link(sale_order.id)],
+            'sale_order_ids': [Command.link(sale_order.id)],
             'advance_payment_method': 'delivered',
         })
         self.assertEqual(sale_order.invoice_status, 'to invoice')
