+++ addons/sale/wizard/sale_make_invoice_advance_views.xml
@@ -2,24 +2,62 @@
 <odoo>
 
     <record id="view_sale_advance_payment_inv" model="ir.ui.view">
-        <field name="inherit_id" ref="account.view_account_advance_payment"/>
-        <field name="mode">primary</field>
-        <field name="model">sale.advance.payment.inv</field>
         <field name="name">Invoice Orders</field>
+        <field name="model">sale.advance.payment.inv</field>
         <field name="arch" type="xml">
-            <form position="attributes">
-                <attribute name="string">Invoice Sales Order</attribute>
+            <form string="Invoice Sales Order">
+                <field name="display_draft_invoice_warning" invisible="1"/>
+                <field name="has_down_payments" invisible="1"/>
+                <div class="alert alert-warning pb-1" role="alert" invisible="not display_draft_invoice_warning">
+                    <p>There are existing <a name="view_draft_invoices" type="object">Draft Invoices</a> for this Sale Order.</p>
+                    <p invisible="advance_payment_method != 'delivered'">
+                        The new invoice will deduct draft invoices linked to this sale order.
+                    </p>
+                </div>
+                <group>
+                    <field name="sale_order_ids" invisible="1"/>
+                    <field name="count" invisible="count == 1"/>
+                    <field name="consolidated_billing" invisible="count == 1"/>
+                    <field name="advance_payment_method" class="oe_inline"
+                        widget="radio"
+                        invisible="count &gt; 1"/>
+                </group>
+                <group name="down_payment_specification"
+                    invisible="advance_payment_method not in ('fixed', 'percentage')">
+                    <field name="company_id" invisible="1"/>
+                    <label for="amount"/>
+                    <div id="payment_method_details">
+                        <field name="currency_id" invisible="1"/>
+                        <field name="display_invoice_amount_warning" invisible="1"/>
+                        <field name="fixed_amount"
+                            invisible="advance_payment_method != 'fixed'"
+                            required="advance_payment_method == 'fixed'"
+                            class="oe_inline"/>
+                        <field name="amount"
+                            invisible="advance_payment_method != 'percentage'"
+                            required="advance_payment_method == 'percentage'"
+                            class="oe_inline"/>
+                        <span invisible="advance_payment_method != 'percentage'"
+                            class="oe_inline">% </span>
+                        <span invisible="not display_invoice_amount_warning"
+                              class="oe_inline text-danger"
+                              title="The Down Payment is greater than the amount remaining to be invoiced.">
+                            <i class="fa fa-warning"/>
+                        </span>
+                    </div>
+                </group>
+                <group invisible="not has_down_payments">
+                    <field name="amount_invoiced"/>
+                    <field name="amount_to_invoice"/>
+                </group>
+                <footer>
+                    <button name="create_invoices" type="object"
+                        id="create_invoice_open"
+                        string="Create Draft"
+                        class="btn-primary" data-hotkey="q"/>
+                    <button string="Cancel" class="btn-secondary" special="cancel" data-hotkey="x"/>
+                </footer>
             </form>
-            <a name="view_draft_invoices" position="inside">There are existing Draft Invoices for this Sale Order.</a>
-            <p name="deduct_draft" position="inside">
-                The new invoice will deduct draft invoices linked to this sale order.
-            </p>
-            <field name="count" position="after">
-                <field name="consolidated_billing" invisible="count == 1"/>
-            </field>
-            <span name="amount_warning" position="attributes">
-                <attribute name="title">The Down Payment is greater than the amount remaining to be invoiced.</attribute>
-            </span>
         </field>
     </record>
 
@@ -27,7 +65,6 @@
         <field name="name">Create invoice(s)</field>
         <field name="res_model">sale.advance.payment.inv</field>
         <field name="view_mode">form</field>
-        <field name="view_id" ref="sale.view_sale_advance_payment_inv"/>
         <field name="target">new</field>
         <field name="binding_model_id" ref="sale.model_sale_order"/>
         <field name="binding_view_types">list</field>
