+++ addons/sale_timesheet/wizard/sale_make_invoice_advance.py
@@ -17,11 +17,11 @@ class SaleAdvancePaymentInv(models.TransientModel):
 
     #=== COMPUTE METHODS ===#
 
-    @api.depends('order_ids')
+    @api.depends('sale_order_ids')
     def _compute_invoicing_timesheet_enabled(self):
         for wizard in self:
             wizard.invoicing_timesheet_enabled = bool(
-                wizard.order_ids.order_line.filtered(
+                wizard.sale_order_ids.order_line.filtered(
                     lambda sol: sol.invoice_status == 'to invoice'
                 ).product_id.filtered(
                     lambda p: p._is_delivered_timesheet()
@@ -46,6 +46,6 @@ class SaleAdvancePaymentInv(models.TransientModel):
             return sale_orders.with_context(
                 timesheet_start_date=self.date_start_invoice_timesheet,
                 timesheet_end_date=self.date_end_invoice_timesheet
-            )._create_invoices(final=True, grouped=not self.consolidated_billing)
+            )._create_invoices(final=self.deduct_down_payments, grouped=not self.consolidated_billing)
 
         return super()._create_invoices(sale_orders)
